pyproject.toml
[project]
name = "tts-papertrail"
description = "Teamtools Studio Reporting Tools."
readme = "README.md"
requires-python = ">=3.6"
license = {text = "Unlicense"}
authors = [
  { name = "muszynsk", email = "matt.muszynski@jpl.nasa.gov" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Unlicense",
]
dependencies = [
    "datetime",
    "lxml",
    "pytest",
    "requests",
    "rich",
    "python-docx",
    "python-pptx",
    "xlsxwriter",
    "tts-html-utils>=0.4.0,<1.0.0",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/NASA-JPL-Teamtools-Studio/papertrail"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[build-system]
requires = [
    "setuptools", 
    "setuptools_scm",
    "toml", 
    "wheel", 
    "tts-utilities>=0.2.1,<1.0.0"
    ]
build-backend = "setuptools.build_meta"

